Managing environments

+++ pkg/dashboard/git/nativeGit/repoCache.go
@@ -14,6 +14,7 @@ import (
 	"github.com/gimlet-io/gimlet-cli/pkg/dashboard/git/customScm"
 	"github.com/gimlet-io/gimlet-cli/pkg/dashboard/git/genericScm"
 	"github.com/gimlet-io/gimlet-cli/pkg/dashboard/server/streaming"
+	"github.com/gimlet-io/gimlet-cli/pkg/git/nativeGit"
 	"github.com/gimlet-io/go-scm/scm"
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/config"
@@ -41,6 +42,8 @@ type RepoCache struct {
 	config      *dashboardConfig.Config
 
 	clientHub *streaming.ClientHub
+
+	lock sync.Mutex
 }
 
 func NewRepoCache(
@@ -119,7 +122,9 @@ func (r *RepoCache) syncGitRepo(repoName string) {
 		return // preventing a race condition in cleanup
 	}
 
-	err = r.repos[repoName].Fetch(&git.FetchOptions{
+	repo := r.repos[repoName]
+
+	err = repo.Fetch(&git.FetchOptions{
 		RefSpecs: fetchRefSpec,
 		Auth: &http.BasicAuth{
 			Username: user,
@@ -133,8 +138,27 @@ func (r *RepoCache) syncGitRepo(repoName string) {
 		return
 	}
 	if err != nil {
-		logrus.Errorf("could not pull: %s", err)
+		logrus.Errorf("could not fetch: %s", err)
 		r.cleanRepo(repoName)
+	}
+
+	w, err := repo.Worktree()
+	if err != nil {
+		logrus.Errorf("could not get working copy: %s", err)
+		r.cleanRepo(repoName)
+		return
+	}
+
+	headBranch := nativeGit.HeadBranch(repo)
+	branchHeadHash := nativeGit.BranchHeadHash(repo, headBranch)
+	err = w.Reset(&git.ResetOptions{
+		Commit: branchHeadHash,
+		Mode:   git.HardReset,
+	})
+	if err != nil {
+		logrus.Errorf("could not reset: %s", err)
+		r.cleanRepo(repoName)
+		return
 	}
 
 	jsonString, _ := json.Marshal(streaming.StaleRepoDataEvent{
@@ -145,24 +169,23 @@ func (r *RepoCache) syncGitRepo(repoName string) {
 }
 
 func (r *RepoCache) cleanRepo(repoName string) {
-	mutex.Lock()
+	r.lock.Lock()
 	delete(r.repos, repoName)
-	mutex.Unlock()
+	r.lock.Unlock()
 }
 
-var mutex = &sync.Mutex{}
-
-func (r *RepoCache) InstanceForRead(repoName string) (*git.Repository, error) {
-	mutex.Lock()
+func (r *RepoCache) InstanceForRead(repoName string) (instance *git.Repository, err error) {
+	r.lock.Lock()
 	if repo, ok := r.repos[repoName]; ok {
-		mutex.Unlock()
-		return repo, nil
+		instance = repo
+	} else {
+		repo, err = r.clone(repoName)
+		instance = repo
+		go r.registerWebhook(repoName)
 	}
-	repo, err := r.clone(repoName)
-	mutex.Unlock()
-	go r.registerWebhook(repoName)
+	r.lock.Unlock()
 
-	return repo, err
+	return instance, err
 }
 
 func (r *RepoCache) InstanceForWrite(repoName string) (*git.Repository, string, error) {
@@ -171,16 +194,15 @@ func (r *RepoCache) InstanceForWrite(repoName string) (*git.Repository, string, 
 		errors.WithMessage(err, "couldn't get temporary directory")
 	}
 
-	mutex.Lock()
+	r.lock.Lock()
 	if _, ok := r.repos[repoName]; !ok {
 		_, err = r.clone(repoName)
-		mutex.Unlock()
-		if err != nil {
-			errors.WithMessage(err, "couldn't clone")
-		}
 		go r.registerWebhook(repoName)
 	}
-	mutex.Unlock()
+	r.lock.Unlock()
+	if err != nil {
+		return nil, "", err
+	}
 
 	repoPath := filepath.Join(r.cachePath, strings.ReplaceAll(repoName, "/", "%"))
 	err = copy.Copy(repoPath, tmpPath)
@@ -205,6 +227,10 @@ func (r *RepoCache) Invalidate(repoName string) {
 }
 
 func (r *RepoCache) clone(repoName string) (*git.Repository, error) {
+	if repoName == "" {
+		return nil, fmt.Errorf("repo name is mandatory")
+	}
+
 	repoPath := filepath.Join(r.cachePath, strings.ReplaceAll(repoName, "/", "%"))
 
 	os.RemoveAll(repoPath)
